Created a super basic folder structure, created makefile for linux builds
